requirements.txt
torch==1.9.1
torchvision==0.10.1
torchaudio===0.9.1
pandas==1.3.3
joblib==1.0.1
scipy==1.7.1
natsort==7.1.1
tqdm==4.62.2
streamlit==1.0.0
tensorflow==2.6.0
matplotlib==3.4.2
plotly==5.3.1
numpy==1.19.5
scikit_image==0.18.1
Pillow==8.3.2
